WPF migration to .NET Core framework
