+++ PSW_Wpf_app/App.xaml
@@ -1,7 +1,7 @@
-﻿<Application x:Class="WpfApp.App"
+﻿<Application x:Class="PSW_Wpf_app.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:WpfApp"
+             xmlns:local="clr-namespace:PSW_Wpf_app"
              StartupUri="MainWindow.xaml">
     <Application.Resources>
