+++ PSW_Wpf_app/BindableBase.cs
