+++ PSW_Wpf_app/Images/parking.jpg
