+++ PSW_Wpf_app/Images/waiting-room.png
