+++ PSW_Wpf_app/Model/GraphicElement.cs
