+++ PSW_Wpf_app/MyICommand.cs
