+++ PSW_Wpf_app/View/ChosenFloorView.xaml.cs
