+++ PSW_Wpf_app/View/MainBuildingView.xaml.cs
