+++ PSW_Wpf_app/pediatrics.txt
@@ -1,0 +1,126 @@
+﻿elipsa,reception,173,87,222,144,0,WORKING_TIME:-10h
+waitingRoom,stuf1,22,10,259,177,0,
+waitingRoom,stuf2,22,10,298,177,0,
+waitingRoom,stuf3,22,10,335,177,0,
+waitingRoom,stuf4,7,10,267,187,0,
+waitingRoom,stuf5,7,10,306,187,0,
+waitingRoom,stuf6,7,10,344,187,0,
+door,door1,9,99,10,155,0,
+door,door2,99,9,296,10,0,
+door,door3,9,99,636,155,0,
+elipsa,table,79,30,60,63,0,
+waitingRoom,chair1,7,62,21,47,0,
+waitingRoom,chair2,15,7,26,102,0,
+waitingRoom,chair3,15,9,26,47,0,
+waitingRoom,chair4,15,15,26,56,0,
+waitingRoom,chair5,15,15,26,71,0,
+waitingRoom,chair6,15,16,26,86,0,
+waitingRoom,chair7,63,7,21,19,0,
+waitingRoom,chair8,10,15,21,26,0,
+waitingRoom,chair9,9,15,75,26,0,
+waitingRoom,chair10,15,15,30,26,0,
+waitingRoom,chair11,15,15,45,26,0,
+waitingRoom,chair12,15,15,60,26,0,
+waitingRoom,chair13,63,7,99,19,0,
+waitingRoom,chair14,10,15,99,26,9,
+waitingRoom,chair15,10,15,152,26,0,
+waitingRoom,chair16,15,15,109,26,0,
+waitingRoom,chair17,15,15,124,26,0,
+waitingRoom,chair18,15,15,139,26,0,
+room,baby_room,97,76,177,10,0,WORKING_TIME:24h
+restroom,toilet1,54,46,450,10,0,
+restroom,toilet2,54,46,502,10,0,
+stairs,stairs,89,83,556,10,0,
+elevator,elevator,56,39,589,93,0,
+door,ulaz7,4,18,589,102,0,
+room,laboratorija,125,140,520,320,0,WORKING_TIME:24h
+room,examination1,142,97,378,363,0,WORKING_TIME:24h
+room,examination2,146,96,232,363,0,WORKING_TIME:24h
+door,door10,42,4,286,363,0,
+room,examination3,140,96,92,363,0,WORKING_TIME:24h
+door,door11,4,37,228,378,0,
+door,door12,37,5,139,363,0,
+room,office,82,179,10,280,0,WORKING_TIME:24h
+waitingRoom,chair19,10,8,222,296,0,
+waitingRoom,chair20,10,8,237,296,0,
+waitingRoom,chair21,10,8,252,296,0,
+waitingRoom,chair22,10,8,267,296,0,
+waitingRoom,chair23,10,8,282,296,0,
+waitingRoom,chair24,10,8,298,296,0,
+waitingRoom,chair25,10,8,313,296,0,
+waitingRoom,chair26,10,8,328,296,0,
+waitingRoom,chair27,10,8,343,296,0,
+waitingRoom,chair28,10,8,358,296,0,
+waitingRoom,chair29,10,8,373,296,0,
+waitingRoom,chair30,10,8,222,309,0,
+waitingRoom,chair31,10,8,237,309,0,
+waitingRoom,chair32,10,8,252,309,0,
+waitingRoom,chair33,10,8,267,309,0,
+waitingRoom,chair34,10,8,282,309,0,
+waitingRoom,chair35,10,8,298,309,0,
+waitingRoom,chair36,10,8,313,309,0,
+waitingRoom,chair37,10,8,328,309,0,
+waitingRoom,chair38,10,8,344,309,0,
+waitingRoom,chair39,10,8,359,309,0,
+waitingRoom,chair40,10,8,373,309,0,
+door,door4,37,4,458,52,0,
+door,door5,37,4,509,52,0,
+door,door6,4,42,177,26,0,
+door,door8,42,4,567,320,0,
+door,door9,42,4,419,363,0,
+door,door13,42,4,26,280,0,
+room,play_area,175,100,10,10,1,WORKING_TIME:24h
+room,canteen,142,81,185,10,1,WORKING_TIME:24h
+door,door1,44,4,70,106,1,
+door,door3,44,4,70,106,1,
+door,door4,43,4,238,87,1,
+restroom,toilet1,54,46,437,10,1,
+restroom,toilet2,54,46,491,10,1,
+stairs,stairs,100,100,545,10,1,
+elevator,elevator,56,39,588,108,1,
+door,door5,26,4,449,52,1,
+door,door6,26,4,507,52,1,
+door,door7,4,18,588,118,1,
+room,neonatal,142,93,10,173,1,
+room,neurology,142,94,10,266,1,WORKING_TIME:24h
+room,rehabilitation,142,100,10,360,1,WORKING_TIME:24h
+door,door8,4,43,148,201,1,
+door,door9,4,43,148,295,1,
+door,door10,4,43,148,382,1,
+room,day_care1,153,100,491,173,1,WORKING_TIME:24h
+room,day_care2,153,100,491,273,1,WORKING_TIME:24h
+room,day_care3,153,87,491,373,1,WORKING_TIME:24h
+door,door12,4,43,491,295,1,
+door,door22,4,43,491,201,1,
+room,reception,13,90,246,201,1,
+room,reception,40,14,246,286,1,
+room,reception,98,13,278,295,1,
+room,reception,40,14,365,287,1,
+room,reception,13,86,391,201,1,
+elipsa,chair1,26,18,266,235,1,
+elipsa,chair2,32,17,309,249,1,
+elipsa,chair3,26,18,356,238,1,
+waitingRoom,chair4,10,8,242,417,1,
+waitingRoom,chair5,10,8,256,417,1,
+waitingRoom,chair6,10,8,271,417,1,
+waitingRoom,chair7,10,8,286,417,1,
+waitingRoom,chair8,10,8,301,417,1,
+waitingRoom,chair9,10,8,317,417,1,
+waitingRoom,chair10,10,8,332,417,1,
+waitingRoom,chair11,10,8,347,417,1,
+waitingRoom,chair12,10,8,362,417,1,
+waitingRoom,chair13,10,8,378,417,1,
+waitingRoom,chair14,10,8,392,417,1,
+waitingRoom,chair15,10,8,242,430,1,
+waitingRoom,chair16,10,8,256,430,1,
+waitingRoom,chair17,10,8,271,430,1,
+waitingRoom,chair18,10,8,286,430,1,
+waitingRoom,chair19,10,8,301,430,1,
+waitingRoom,chair20,10,8,317,430,1,
+waitingRoom,chair21,10,8,332,430,1,
+waitingRoom,chair22,10,8,347,430,1,
+waitingRoom,chair4,10,8,363,430,1,
+waitingRoom,chair4,10,8,378,430,1,
+waitingRoom,chair4,10,8,392,430,1,
+door,door23,4,43,185,32,1,
+door,door24,4,43,491,395,1,
