+++ WellDevCore.sln
@@ -1,13 +1,13 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.1062
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.30011.22
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WellDevCore", "WellDevCore\WellDevCore.csproj", "{CA1C100A-1BCE-4743-9D43-4BD4153E792C}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PSW_Web_app", "PSW_Web_app\PSW_Web_app.csproj", "{95FFFA54-4D60-4D5A-A59F-D980E48A94B3}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WpfApp", "WpfApp\WpfApp.csproj", "{A0DFCE36-7E78-4D0E-853D-D5B673BDD3EA}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PSW_Wpf_app", "PSW_Wpf_app\PSW_Wpf_app.csproj", "{12D04F2A-8928-4F8F-BA16-D8AA6B36091A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -23,10 +23,10 @@ Global
 		{95FFFA54-4D60-4D5A-A59F-D980E48A94B3}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{95FFFA54-4D60-4D5A-A59F-D980E48A94B3}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{95FFFA54-4D60-4D5A-A59F-D980E48A94B3}.Release|Any CPU.Build.0 = Release|Any CPU
-		{A0DFCE36-7E78-4D0E-853D-D5B673BDD3EA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A0DFCE36-7E78-4D0E-853D-D5B673BDD3EA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A0DFCE36-7E78-4D0E-853D-D5B673BDD3EA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A0DFCE36-7E78-4D0E-853D-D5B673BDD3EA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{12D04F2A-8928-4F8F-BA16-D8AA6B36091A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{12D04F2A-8928-4F8F-BA16-D8AA6B36091A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{12D04F2A-8928-4F8F-BA16-D8AA6B36091A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{12D04F2A-8928-4F8F-BA16-D8AA6B36091A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
