--- WpfApp/MainWindow.xaml
@@ -1,32 +1,0 @@
-﻿<Window x:Class="WpfApp.MainWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:viewmodel="clr-namespace:WpfApp.ViewModel" 
-        xmlns:local="clr-namespace:WpfApp"
-        mc:Ignorable="d"
-        Title="MainWindow" Height="500" Width="900">
-    <Window.Resources>
-        <viewmodel:ShapeViewModel x:Key="vm"></viewmodel:ShapeViewModel>
-    </Window.Resources>
-    <Grid Background="White"  DataContext="{Binding Source={StaticResource vm}}">
-        <!--<Canvas Background="LightGreen" HorizontalAlignment="Left" Height="470" Margin="0" VerticalAlignment="Top" Width="655" Name="Can1"/>-->
-        <UserControl Name="userControl" Margin="0,0,224.6,1.4"></UserControl>
-        <Grid Background="#FFAAC4C5" HorizontalAlignment="Left" Height="469" Margin="674,0,0,0" VerticalAlignment="Top" Width="218">
-            <Grid Background="CadetBlue" HorizontalAlignment="Left" Height="370" Margin="10,55,0,0" VerticalAlignment="Top" Width="198">
-                <TextBlock HorizontalAlignment="Left" Height="24" Margin="10,10,0,0" TextWrapping="Wrap" Text="Choose building:" FontSize="16" VerticalAlignment="Top" Width="150"/>
-                <ComboBox HorizontalAlignment="Left" Height="24" Margin="10,42,0,0" VerticalAlignment="Top" Width="181" ItemsSource="{Binding Path=Buildings}" Name="buildC"  DisplayMemberPath="Name"/>
-                <TextBlock HorizontalAlignment="Left" Height="24" Margin="10,88,0,0" TextWrapping="Wrap" Text="Choose the floor:" FontSize="16" VerticalAlignment="Top" Width="150"/>
-                <ComboBox HorizontalAlignment="Left" Height="24" Margin="10,117,0,0" VerticalAlignment="Top" Width="181"  ItemsSource="{Binding Path=FloorsCount}" Name="floorC"/>
-                <Button Content="Apply" HorizontalAlignment="Center" Height="42" Margin="10,172,7,0" VerticalAlignment="Top" Width="181" FontSize="28" FontFamily="Segoe UI Semibold" Click="Button_Click"/>
-            </Grid>
-            <Button FontSize="12" FontFamily="Segoe UI Semibold"  HorizontalAlignment="Left" Height="28" Margin="10,16,0,0" VerticalAlignment="Top" Width="40" Click="Button_Click_HOME">
-                <Image Source="Images/home.png"></Image>
-            </Button>
-            <Button Content="Map Information" FontSize="12" FontFamily="Segoe UI Semibold"  HorizontalAlignment="Left" Height="28" Margin="108,16,0,0" VerticalAlignment="Top" Width="100" Click="Button_Click_HOME" IsEnabled="False"/>
-
-        </Grid>
-    </Grid>
-
-</Window>
