--- WpfApp/WpfApp.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30011.22
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WpfApp", "WpfApp.csproj", "{A0DFCE36-7E78-4D0E-853D-D5B673BDD3EA}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A0DFCE36-7E78-4D0E-853D-D5B673BDD3EA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{A0DFCE36-7E78-4D0E-853D-D5B673BDD3EA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{A0DFCE36-7E78-4D0E-853D-D5B673BDD3EA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{A0DFCE36-7E78-4D0E-853D-D5B673BDD3EA}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {4ED21506-22E8-49F1-B430-F0FABD3ABDDD}
-	EndGlobalSection
-EndGlobal
