--- WpfApp/externalMap.txt
@@ -1,32 +1,0 @@
-Building,194,157,62,61,Surgical,This building has a ground floor and a first floor. In the ground floor there are ambulance and ER.Patience can come here when they have an unexcepted injury or fractue.We have an excellent X-ray and many other medical devices in our diagnostic office. On the ground floor there are also info desk and waiting room. On the first floor there are operating rooms and recovering rooms. We also have elevators for every entrance which patients can use.info pult->contact +0044/014789889
-Building,186,157,408,61,MedicalCenter,This building has a ground floor and 3 more floors. On the ground floor there are info desk and cardiologist and dermatology and ORL and orthopedic and urology and physiatry and endocrinology. On the first floor there are stomatolog and gynecology and physiatry. On the second floor there are general practise and cardiologist and dermatology and ORL . On the third floor there are immunologist and infectology and radiology. info pult -> contact +0044/014678777
-Building,181,141,243,272,Pediatrics,This building has a ground floor and a first floor. On the ground floor there are examination rooms and laboratory and feeding area for babys and stuff room. On the first floor there are pediatrics day care and play area for childreen and pediatric rehabilitation and pediatric neurology. Also there is cantine for lunch and information pult. We also have elevators for every entrance which patients can use.  info pult -> contact 0044/014678543
-Parking,158,141,62,272,,
-Parking,148,141,446,272,,
-Street,33,436,612,24,,
-Street,621,33,10,427,,
-Street,32,418,10,24,,
-Street,635,33,10,10,,
-Street,26,61,317,43,,
-Street,38,26,266,109,,
-Street,49,25,349,110,,
-Street,27,88,317,140,,
-Street,571,26,41,228,,
-Street,26,8,317,254,,
-Street,36,18,484,254,,
-Street,18,33,594,317,,
-Street,36,13,484,414,,
-Street,12,36,434,314,,
-Street,36,17,136,255,,
-Street,13,36,220,314,,
-Street,33,14,125,413,,
-Street,21,33,41,317,,
-Entrance,10,36,233,314,,
-Entrance,46,13,308,259,,
-Entrance,10,36,424,314,,
-Entrance,36,10,484,218,,
-Entrance,10,40,256,100,,
-Entrance,10,40,398,100,,
-Entrance,36,10,136,218,,
-Fountain,60,58,299,92,,
-Circle,27,25,317,110,,
