--- WpfApp/surgicalBranchesFloors.txt
@@ -1,40 +1,0 @@
-room,orthopedic_cabinet,222,83,92,38,0,WorkingTime:24,
-room,diagnostic_cabinet,149,83,92,121,0,WorkingTime:12,
-room,emergency_room_1,149,117,92,204,0,WorkingTime:10,
-room,emergency_room_2,149,117,92,321,0,WorkingTime:8,
-room,emergency_room_3,175,108,402,221,0,WorkingTime:24,
-storageRoom,medical_supplies,175,111,402,327,0,WorkingTime:24,
-counter,info_pult,75,119,280, 161,0,WorkingTime:24,
-restroom,F_restroom,75,60,314,38,0,,
-restroom,M_restroom,75,60,389,38,0,,
-elevator,elevator_bottom,83,37,402,402,0,,
-elevator,elevator_top,57,83,463,38,0,,
-stairs,stairs, 57,83,520,38,0,,
-door,entrance_1,10,60,567,144,0,,
-door,entrance_2,60,10,295,428,0,,
-door,1,40,6,147,115,0,,
-door,2,40,6,266,115,0,,
-door,3,40,6,328,92,0,,
-door,4,40,6,389,38,0,,
-door,5,6,40,235,139,0,,
-door,6,6,40,235,240,0,,
-door,7,6,40,235,360,0,,
-door,8,6,40,402,240,0,,
-door,9,6,40,402,332,0,,
-door,10,40,6,407,92,0,,
-waitingRoom,watingRoom1,60,60,402,144,0,,
-restroom,F_restroom,75,60,314,38,0,,
-waitingRoom,watingRoom2,60,60,290,303,0,,
-room,preoperative,73,200,241,121,1,WorkingTime:24,
-room,operating_room,222,83,92,38,1,WorkingTime:24,
-room,operating_room,149,83,92,121,1,WorkingTime:24,
-room,operating_room,149,117,92,204,1,WorkingTime:24,
-room,intensive_care,222,117,92,321,1,WorkingTime:24,
-room,recovery_room_1,175,83,402,177,1,WorkingTime:24,
-room,recovery_room_2,175,67,402,260,1,WorkingTime:24,
-room,recovery_room_3,175,111,402,327,1,WorkingTime:24,
-restroom,F_restroom,75,60,314,38,1,,
-restroom,M_restroom,75,60,389,38,1,,
-elevator,elevator_bottom,83,37,402,402,1,,
-elevator,elevator_top,57,83,463,38,1,,
-stairs,stairs,57,83, 520,38,1,,
